+++ actions/executions_cancel.py
@@ -1,21 +1,20 @@
 from lib.action import St2BaseAction
-from lib.formatters import format_result
 
 __all__ = [
-    'St2ExecutionsResume'
+    'St2ExecutionsCancel'
 ]
 
 
-class St2ExecutionsResume(St2BaseAction):
+class St2ExecutionsCancel(St2BaseAction):
     def run(self, ids):
         result = {}
         success = True
         for i in ids:
             try:
-                res = self.client.liveactions.resume(execution_id=i)
+                res = self.client.executions.delete_by_id(instance_id=i)
             except Exception as exc:
                 result[i] = str(exc)
                 success = False
             else:
-                result[i] = format_result(item=res)
+                result[i] = res
         return success, result